Refactor timing utilities into shared module
